refactor: standardize definition of 'author' contributors
and replace redis caching of codebase all_contributor lists with
querysets (did save a few queries but doesn't seem to have any
meaningful performance impact)

codebases were considering any citable release contributor as an author
and releases considered anyone with a role of "author" to be an author.
Now we use a union of the two -- not sure if this is the best way but
regardless, its easier to change since it all stems from authors()
and nonauthors() on the ReleaseContributorQuerySet

codebase and release both now have 'nonauthor contributor' accessors,
which is useful because this is what things like codemeta/datacite/etc.
consider 'contributors'