feat(WIP): add release FS API for imported releases
refactors the `CodebaseReleaseFsApi` to inherit from an abstract
`BaseCodebaseReleaseFsApi`, along with the new
`ImportedCodebaseReleaseFsApi`

the main function of the imported release fs api is to import releases
from a remote source by downloading an archive to originals/, extracting
to sip/data/ and then using the inherited functionality to build
archives for review/publishing. It does not implement methods for
dealing with files directly outside of this importing workflow like
add(), delete(), rebuild(), etc.

the imported release fs api manages a manifest for keeping track of file
categories. Using this for all releases is potentially something to do
in the future but is a rather complicated refactor that had to be dialed
back here